MainActivity - reformatting for different screen
I separated the login screen from the MainActivity.

TODO: Separate view from viewmodel of the login
