+++ app/src/main/java/com/lastaoutdoor/lasta/view/MainActivity.kt
@@ -1,5 +1,6 @@
 package com.lastaoutdoor.lasta.view
 
+import android.app.Activity
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.rememberLauncherForActivityResult
@@ -35,32 +36,22 @@ import com.firebase.ui.auth.FirebaseAuthUIActivityResultContract
 import com.firebase.ui.auth.data.model.FirebaseAuthUIAuthenticationResult
 import com.google.firebase.auth.FirebaseAuth
 import com.lastaoutdoor.lasta.R
+import com.lastaoutdoor.lasta.view.screen.LoginScreen
+import com.lastaoutdoor.lasta.view.screen.MapScreen
 import com.lastaoutdoor.lasta.view.theme.LastaTheme
 
 class MainActivity : ComponentActivity() {
   override fun onCreate(savedInstanceState: Bundle?) {
     super.onCreate(savedInstanceState)
-    setContent { GlobalNav() }
-  }
+    setContent {
 
-  private fun onSignInResult(
-      result: FirebaseAuthUIAuthenticationResult,
-      onNavigateToMain: () -> Unit
-  ) {
-    val response = result.idpResponse
-    if (result.resultCode == RESULT_OK) {
-      // Successfully signed in
-      val user = FirebaseAuth.getInstance().currentUser
-      onNavigateToMain()
-    }
-  }
+      //The navcontroller that allows to go from the login screen to the main menu screen
+      val navController = rememberNavController()
+      NavHost(navController = navController, startDestination = "login") {
+        composable("mainMenu") { MainMenu() }
+        composable("login") { LoginScreen(onNavigateToMain = { navController.navigate("mainMenu") }) }
+      }
 
-  @Composable
-  fun GlobalNav() {
-    val navController = rememberNavController()
-    NavHost(navController = navController, startDestination = "login") {
-      composable("mainMenu") { MainMenu() }
-      composable("login") { LoginMenu(onNavigateToMain = { navController.navigate("mainMenu") }) }
     }
   }
 
@@ -73,62 +64,7 @@ class MainActivity : ComponentActivity() {
 
   @Composable
   fun MapTab() {
-    Text(text = "map tab")
+    MapScreen();
   }
 
-  @Composable
-  fun LoginMenu(onNavigateToMain: () -> Unit) {
-    val providers =
-        arrayListOf(
-            AuthUI.IdpConfig.GoogleBuilder().build(),
-        )
-    // Create and launch sign-in intent
-    val signInIntent =
-        AuthUI.getInstance().createSignInIntentBuilder().setAvailableProviders(providers).build()
-    val signInLauncher =
-        rememberLauncherForActivityResult(
-            FirebaseAuthUIActivityResultContract(),
-        ) { r ->
-          this.onSignInResult(r, onNavigateToMain)
-        }
-    LastaTheme {
-      Column(
-          modifier = Modifier.fillMaxSize().padding(15.dp),
-          verticalArrangement = Arrangement.Center,
-          horizontalAlignment = Alignment.CenterHorizontally) {
-            Image(
-                painter = painterResource(id = R.drawable.app_logo),
-                contentDescription = "App Logo",
-                modifier = Modifier.size(189.dp))
-            Spacer(modifier = Modifier.size(35.dp))
-            Text(
-                text = "LASTA",
-                modifier = Modifier.width(256.dp).height(65.dp),
-                style =
-                    TextStyle(
-                        fontSize = 57.sp,
-                        lineHeight = 64.sp,
-                        color = Color(0xFF191C1E),
-                        textAlign = TextAlign.Center,
-                    ))
-            Spacer(modifier = Modifier.size(150.dp))
-            Button(
-                onClick = { signInLauncher.launch(signInIntent) },
-                shape = RoundedCornerShape(30.dp),
-                colors = ButtonDefaults.buttonColors(containerColor = Color.White),
-                border = BorderStroke(1.dp, Color.Gray),
-                content = {
-                  Image(
-                      painter = painterResource(id = R.drawable.google_logo),
-                      contentDescription = "Google Logo",
-                      modifier = Modifier.size(24.dp))
-                  Text(
-                      text = "Sign in with Google",
-                      modifier = Modifier.padding(6.dp),
-                      color = Color.Black,
-                      fontSize = 14.sp)
-                })
-          }
-    }
-  }
 }
