+++ app/src/main/java/com/lastaoutdoor/lasta/view/screen/MapScreen.kt
@@ -1,0 +1,13 @@
+package com.lastaoutdoor.lasta.view.screen
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.material3.Text
+
+import androidx.compose.runtime.Composable
+
+@Composable
+fun MapScreen() {
+    Column {
+        Text("Map Screen")
+    }
+}
