+++ app/src/main/java/com/lastaoutdoor/lasta/viewmodel/dummyFIle.md
@@ -1,0 +1,1 @@
+Just added this file temporarily to be able to push it
