+++ app/src/test/java/com/lastaoutdoor/lasta/screen/SecondScreen.kt
@@ -1,14 +1,18 @@
 package com.lastaoutdoor.lasta.screen
 
+
+//This file is kepts as a reference for testing, but this actual screens has been deleted
+
+
 import androidx.compose.ui.test.SemanticsNodeInteractionsProvider
-import com.lastaoutdoor.lasta.resources.C
+//import com.lastaoutdoor.lasta.resources.C
 import io.github.kakaocup.compose.node.element.ComposeScreen
 import io.github.kakaocup.compose.node.element.KNode
 
-class SecondScreen(semanticsProvider: SemanticsNodeInteractionsProvider) :
-    ComposeScreen<SecondScreen>(
-        semanticsProvider = semanticsProvider,
-        viewBuilderAction = { hasTestTag(C.Tag.second_screen_container) }) {
-
-  val simpleText: KNode = child { hasTestTag(C.Tag.greeting_robo) }
-}
+//class SecondScreen(semanticsProvider: SemanticsNodeInteractionsProvider) :
+//    ComposeScreen<SecondScreen>(
+//        semanticsProvider = semanticsProvider,
+//        viewBuilderAction = { hasTestTag(C.Tag.second_screen_container) }) {
+//
+//  val simpleText: KNode = child { hasTestTag(C.Tag.greeting_robo) }
+//}
